refactor: extract ScrollingName, AnimatedNumber, and MetricsSkeleton to UserPerformanceCard; update imports in campaign-analytics section
