+++ src/components/campaign-analytics/index.js
@@ -2,3 +2,9 @@
 export { default as PlatformOverviewMobile } from './PlatformOverviewMobile';
 export { default as ManualCreatorEntryForm } from './ManualCreatorEntryForm';
 export { default as PlatformOverviewDesktop } from './PlatformOverviewDesktop';
+export {
+  ScrollingName,
+  AnimatedNumber,
+  MetricsSkeleton,
+  default as UserPerformanceCard,
+} from './UserPerformanceCard';
